fixing some issues
