+++ .trunk/configs/.cspell.json
@@ -122,6 +122,7 @@
     "Doctest",
     "Doctests",
     "linenums",
-    "XLYOFNOQVPJJNP"
+    "XLYOFNOQVPJJNP",
+    "CRIPTUUID"
   ]
 }
