+++ src/cript/nodes/uuid_base.py
@@ -7,10 +7,6 @@ from cript.nodes.core import BaseNode
 from cript.nodes.exceptions import CRIPTUUIDException
 
 
-def get_uuid_from_uid(uid):
-    return str(uuid.UUID(uid[2:]))
-
-
 class UUIDBaseNode(BaseNode, ABC):
     """
     Base node that handles UUIDs and URLs.
@@ -33,7 +29,7 @@ class UUIDBaseNode(BaseNode, ABC):
 
     _json_attrs: JsonAttributes = JsonAttributes()
 
-    def __new__(cls, **kwargs):
+    def __new__(cls, *args, **kwargs):
         uuid: Optional[str] = kwargs.get("uuid")
         if uuid and uuid in UUIDBaseNode._uuid_cache:
             existing_node_to_overwrite = UUIDBaseNode._uuid_cache[uuid]
@@ -44,6 +40,8 @@ class UUIDBaseNode(BaseNode, ABC):
         return new_uuid_node
 
     def __init__(self, **kwargs):
+        from cript.nodes.util.core import get_uuid_from_uid
+
         # initialize Base class with node
         super().__init__(**kwargs)
         # Respect uuid if passed as argument, otherwise construct uuid from uid
@@ -62,12 +60,6 @@ class UUIDBaseNode(BaseNode, ABC):
 
         api = _get_global_cached_api()
         return f"{api.host}/{api.api_prefix}/{api.api_version}/{self.uuid}"
-
-    def __deepcopy__(self, memo):
-        node = super().__deepcopy__(memo)
-        node._json_attrs = replace(node._json_attrs, uuid=get_uuid_from_uid(node.uid))
-        UUIDBaseNode._uuid_cache[str(node.uuid)] = node
-        return node
 
     @property
     def updated_by(self):
